Reset the time.
It may be nice to be able to reset the time on a cronjob. In order to do
this we need to assert that the time is a CronTime, assert that the
existing job has been stopped (assume that the stop function is
working), and then change the time. We leave it up to the user to
restart the job.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>