Artifially bump only cron style times. (GH-43)
We were artificially bumping all crontimes by 1 second to ensure that
the timeout didn't fire on repetitive jobs many times (omg fire
now...and again, and again and so on...). This was incorrect and
unexpected behavior for specific date/time values. We now only bump cron
style date/time values.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>