fix: cron should still execute after changing the time back during daylight savings (#966)
<!--- Provide a general summary of your changes in the Title above
(following the Conventional Commits standard) -->
<!-- More infos: https://www.conventionalcommits.org -->
<!-- Commit types:
https://github.com/insurgent-lab/conventional-changelog-preset#commit-types-->

## Description

<!--- Describe your changes in detail -->
This PR will fix the issue where cron did not execute after changing the
time back during daylight savings. It also replaces `ts-jest` with
`@swc/jest` which is a little faster at running tests.

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->
fixes #881

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
I ended up doing a big refactor to fix this issue because everything to
do with time zones seemed too complicated to understand as is. I decided
that rather than trying to decide based on what the time "should" be
after a change compared to what it actually was from Luxon, we should
just check if the UTC offset changed. That would indicate a daylight
savings jump with very little work on our part. Then we could just check
if a cron execution was supposed to happen in the gap we skipped.

Building on that idea, I decided to ultimately create a new fake cron
time in UTC based on the timezoned time, do all the manipulations
without worrying about time zones, and then convert back into the
timezoned time and figure out any adjustments if needed.

As for #881 specifically, it looks like some of our date creations were
ambiguous. According to the Luxon docs, [you can sometimes create
ambiguous
times](https://moment.github.io/luxon/#/zones?id=dst-weirdness) if you
try to initialize a `DateTime` object in the middle of a daylight
savings jumping back event. if the time jumps back and hour at 3 am and
I create a `DateTime` at 2:30 am on that day, is it the first 2:30 am or
the second 2:30 am? I switched some of our `DateTime` creations to UTC
to make the time they represented unambiguous, and I added some checks
to see if the previous hour was ambiguous or not before returning the
time.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
Added a new test case, made sure previous test cases passed (and edited
some to make them more clear as to the times being tested), and removed
some test cases that tested functions I removed that we no longer
needed.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [ ] If my change introduces a breaking change, I have added a `!`
after the type/scope in the title (see the Conventional Commits
standard).

---------

Co-authored-by: Pierre Cavin <me@sherlox.io>